pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-test-data-generator"
version = "0.1.1"
description = "Generate structured, customizable test data for API testing."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = " Libin Samkutty" }]
keywords = ["testing", "api", "data-generation", "faker"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "faker>=24.0.0",
    "pydantic>=2.6.0",
    "jsonschema>=4.21.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
csv = ["pandas>=2.2.0"]
cli = ["rich>=13.7.0"]
all = ["pandas>=2.2.0", "rich>=13.7.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
]

[project.scripts]
api-gen = "api_test_data_generator.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["api_test_data_generator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=api_test_data_generator --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["api_test_data_generator"]
omit = ["tests/*"]
